Adds `useCollapseData` prop
This commit adds a new prop allowing a user to manually specify the collapse status of the entire tree on initial load, by setting the _collapse property in their dataset. If useCollapseData is set to false, tree behavior is unchanged. If set to true, the component uses the collapse property set in the data prop. (#114)